feat: в core добавлена схема для получения списка элементовв
